workflow-backend/requirements.txt
django==4.2.17
djangorestframework==3.15.2
mongoengine==0.29.1
django-cors-headers==4.6.0
python-dotenv==1.0.1
